[feature/hexagonal-architecture-for-repo] adding ProductDiscount feature

+++ src/main/java/com/inditex/priceextractor/domain/PositiveNumber.java
@@ -2,7 +2,7 @@ package com.inditex.priceextractor.domain;
 
 import com.inditex.priceextractor.domain.exception.PositiveNumberException;
 
-public record PositiveNumber(int value) {
+public record PositiveNumber(Double value) {
 
   public PositiveNumber {
     if (value < 0) {
